Move Bounds to model
