+++ app/src/main/java/org/groundplatform/android/persistence/local/room/converter/ConverterExt.kt
@@ -31,6 +31,7 @@ import org.groundplatform.android.model.job.Job
 import org.groundplatform.android.model.job.Job.DataCollectionStrategy
 import org.groundplatform.android.model.job.Style
 import org.groundplatform.android.model.locationofinterest.LocationOfInterest
+import org.groundplatform.android.model.map.Bounds
 import org.groundplatform.android.model.mutation.LocationOfInterestMutation
 import org.groundplatform.android.model.mutation.SubmissionMutation
 import org.groundplatform.android.model.submission.DraftSubmission
@@ -74,7 +75,6 @@ import org.groundplatform.android.persistence.local.room.relations.JobEntityAndR
 import org.groundplatform.android.persistence.local.room.relations.SurveyEntityAndRelations
 import org.groundplatform.android.persistence.local.room.relations.TaskEntityAndRelations
 import org.groundplatform.android.proto.Survey.DataSharingTerms
-import org.groundplatform.android.ui.map.Bounds
 import org.json.JSONObject
 import timber.log.Timber
