+++ app/src/main/java/org/groundplatform/android/repository/LocationOfInterestRepository.kt
@@ -29,6 +29,7 @@ import org.groundplatform.android.model.geometry.Geometry
 import org.groundplatform.android.model.job.Job
 import org.groundplatform.android.model.locationofinterest.LocationOfInterest
 import org.groundplatform.android.model.locationofinterest.generateProperties
+import org.groundplatform.android.model.map.Bounds
 import org.groundplatform.android.model.mutation.LocationOfInterestMutation
 import org.groundplatform.android.model.mutation.Mutation
 import org.groundplatform.android.model.mutation.Mutation.SyncStatus
@@ -39,7 +40,6 @@ import org.groundplatform.android.persistence.sync.MutationSyncWorkManager
 import org.groundplatform.android.persistence.uuid.OfflineUuidGenerator
 import org.groundplatform.android.proto.Survey.DataVisibility
 import org.groundplatform.android.system.auth.AuthenticationManager
-import org.groundplatform.android.ui.map.Bounds
 import org.groundplatform.android.ui.map.gms.GmsExt.contains
 import timber.log.Timber
