+++ app/src/main/java/org/groundplatform/android/ui/map/MapFragment.kt
@@ -20,6 +20,7 @@ import androidx.annotation.IdRes
 import kotlinx.coroutines.flow.SharedFlow
 import org.groundplatform.android.model.geometry.Coordinates
 import org.groundplatform.android.model.imagery.TileSource
+import org.groundplatform.android.model.map.Bounds
 import org.groundplatform.android.ui.common.AbstractFragment
 
 /** Implementation of Fragment which supports displaying a map. */
