+++ app/src/main/java/org/groundplatform/android/ui/map/gms/mog/TileCoordinates.kt
@@ -19,7 +19,7 @@ package org.groundplatform.android.ui.map.gms.mog
 import kotlin.math.ln
 import kotlin.math.tan
 import org.groundplatform.android.model.geometry.Coordinates
-import org.groundplatform.android.ui.map.Bounds
+import org.groundplatform.android.model.map.Bounds
 
 /**
  * Uniquely identifies the coordinates of a web mercator tile by its X and Y coordinates and its
