+++ app/src/test/java/org/groundplatform/android/persistence/local/LocalOfflineAreaStore.kt
@@ -21,8 +21,8 @@ import dagger.hilt.android.testing.HiltAndroidTest
 import javax.inject.Inject
 import org.groundplatform.android.BaseHiltTest
 import org.groundplatform.android.model.imagery.OfflineArea
+import org.groundplatform.android.model.map.Bounds
 import org.groundplatform.android.persistence.local.stores.LocalOfflineAreaStore
-import org.groundplatform.android.ui.map.Bounds
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
