+++ app/src/test/java/org/groundplatform/android/system/GeocodingManagerTest.kt
@@ -24,7 +24,7 @@ import javax.inject.Inject
 import kotlin.test.assertEquals
 import org.groundplatform.android.BaseHiltTest
 import org.groundplatform.android.model.geometry.Coordinates
-import org.groundplatform.android.ui.map.Bounds
+import org.groundplatform.android.model.map.Bounds
 import org.groundplatform.android.ui.map.gms.GmsExt.center
 import org.junit.Test
 import org.junit.runner.RunWith
