+++ app/src/test/java/org/groundplatform/android/ui/map/CameraPositionTest.kt
@@ -17,6 +17,7 @@ package org.groundplatform.android.ui.map
 
 import com.google.common.truth.Truth.assertThat
 import org.groundplatform.android.FakeData
+import org.groundplatform.android.model.map.Bounds
 import org.junit.Test
 
 class CameraPositionTest {
